+++ packages/website/src/activity-patterns/components/activity-patterns-by-location/activity-patterns-by-location.ts
@@ -2,7 +2,7 @@ import { Options, Vue } from 'vue-class-component'
 import { Prop } from 'vue-property-decorator'
 
 import { generateDetectionHtmlPopup } from '@/activity-patterns/components/activity-patterns-by-location/functions'
-import { ACTIVITY_PATTERN_KEYS } from '@/activity-patterns/functions'
+import { ACTIVITY_PATTERN_MAP_KEYS } from '@/activity-patterns/functions'
 import { getExportFilterName } from '~/dataset-filters/functions'
 import { DEFAULT_LATITUDE, DEFAULT_LONGITUDE } from '~/maps'
 import { MapBubbleComponent, MapDataSet } from '~/maps/map-bubble'
@@ -25,11 +25,11 @@ const DEFAULT_PREFIX = 'Patterns-By-Site'
 export default class ActivityPatternsByLocation extends Vue {
   @Prop({ default: [] }) public datasets!: MapDataSet[]
 
-  selectedDetectionType = ACTIVITY_PATTERN_KEYS.detectionFrequency
-  occupancyType = ACTIVITY_PATTERN_KEYS.occupancy
+  selectedDetectionType = ACTIVITY_PATTERN_MAP_KEYS.detectionFrequency
+  occupancyType = ACTIVITY_PATTERN_MAP_KEYS.occupancy
   datasetTypes: DatasetType[] = [
-    { label: 'Detection', value: ACTIVITY_PATTERN_KEYS.detection },
-    { label: 'Detection frequency', value: ACTIVITY_PATTERN_KEYS.detectionFrequency }
+    { label: 'Detection', value: ACTIVITY_PATTERN_MAP_KEYS.detection },
+    { label: 'Detection frequency', value: ACTIVITY_PATTERN_MAP_KEYS.detectionFrequency }
   ]
 
   isShowLabels = true