+++ packages/website/src/activity-patterns/types.ts
@@ -1,3 +1,5 @@
+import { ActivityPatternsDataByTime } from '~/api/activity-patterns-service'
+
 export interface MetricsDataset {
   percentage: string
   description: string
@@ -8,4 +10,9 @@ export interface Metrics {
   title: string
   information: string
   datasets: MetricsDataset[]
+}
+
+export interface TimeDataset {
+  color: string
+  data: ActivityPatternsDataByTime
 }